ebic
- fixed a bug where the chunks where refreshing very fast (didnt check if the chunks important for the logic which caused the script to run anyways on every chunk even if that chunk got refreshed already)

+++ KitsuneNoMori/Assets/Scripts/ChunkHandling.cs
@@ -142,14 +142,14 @@ public class ChunkHandling : MonoBehaviour
     {
         foreach(ChunkModel chunk in chunks)
         {
-            Debug.Log("searching chunk...");
             if (chunk.ChunkIdentifier == touchedChunk)
             {
-                Debug.Log("Chunk found!");
-                chunks.Where(x => x == chunk).FirstOrDefault().IsUnlocked = true;
-                DeleteDrawables();
-                RefreshDrawables();
-
+                if(chunk.IsUnlocked == false)
+                {
+                    chunks.Where(x => x == chunk).FirstOrDefault().IsUnlocked = true;
+                    DeleteDrawables();
+                    RefreshDrawables();
+                }
             }
         }
     }
